Support for thread suspension
The run queue was formerly implemented as a block RAM indexed by
thread id, with each entry storing the program counter for that
thread. Now it is a proper queue: threads can be removed and put
back in a different order, and instructions can in principle take
any amount of time to complete.  This means that threads will
effectively suspend when they block, e.g. wait for a
load-response from off chip memory, for a timer to complete, or
for a message to arrive.  This change does not affect Fmax, and
has only a tiny impact on area.