Tweaks to ReliableLink
In some cases, the ethernet type/length field can be interpreted by
the MAC in such a way as to prevent normal packet transmission.  So we
now ensure that type-length field is set appropriately.

In addition, underflow errors can occur on the client->MAC stream when
the valid signal is deasserted during packet transmission.  Our use of
a half-rate FIFO meant that this would happen often in practice, so we
now use a full-rate FIFO which solves the problem.