Handle overflow inside the reliable link
Currently, when a packet arrives on the MAC, and there is no space to
store it, I let the MAC drop the packet - the reliable links will
ensure a resend. But this is a bad idea because those dropped packets
contain valuable information - the acks for packets sent in the
opposite direction on the same link.  This means that if the receive
lane backs up, progress is also blocked on the send lane.

This fix moves packet dropping into the reliable link itself,
because here I can observe acks for the transmit lane.

On thing still to explore: is there now too much unnecessary buffering
on the QSys clock-crossing FIFO from the MAC?  Probably -- it's a
waste of BRAMs.