Fix bug in boot loader
The boot loader has support for writing to data memory, and uses a
cache flush to publish these writes before triggering the application.
To ensure that the flush has completed, we always issued a load to a
predefined address.  Problem is, that predefined address (or the cache
line that contains it) could be written to by a boot loader on another
core leading to a race.  I had never observed this bug actually being
trigged until implementing the hardware multicast work, which improves
efficiency of message delivery to the boot loader.  Thanks to
@jrbeaumont for helping to track this down.