Prevent optimising of writes to mailbox scratchpad
Problem: messages are often written to the mailbox scratchpad,
a region of memory that is never read by software (only hardware).

Previous solution: make the pointer to the message slot volatile.
The downside is that ugly "volatile" tags spread into high-level
user code.

New solution: I believe the simple change in this commit removes
the need for volatile.  It basically tells the compiler that a call to
tinselSend may read the message slot.