Faster mcast table construction
There is still a lot of room for improvement here.  One possibility is
to put an upper bound on the size of receiver groups (splitting
these groups to meet the bound), which would allow faster allocation of
routing table entries. Free space in each table could be maintained
using a bit-vector, and the alloctor could OR the free-space tables of
the mcast targets before locating the first 0.

But even for large highly-connected graphs, the time spent constructing
routing tables is now roughly the same as the partition/placement time.
So, for a significant improvement, both of these problems would need
to be tackled.