Better handling of IND records
There was a cyclic dependency in the ProgRouter between trying to send
a IND-routed flit, and trying to consume one which would surely lead
to deadlock in some cases.

This is solved by (1) removing the loopback route around the
ProgRouter, and handling IND records within each fetcher, entirely
independently of the other fetchers; and (2) restricting the number of
IND records allowed per key lookup to one.  For efficiency, we
introduce a second restriction: IND records are only allowed in
max-sized key lookups.  This means that fetchers only block in the
worst case (and even then we allow two outstanding indirections at a
time, per fetcher, to keep utilisation high).  These restrictions
still permit the use case for which IND records are intended:
overcoming the max beat limit of a key lookup.

While doing this I also spotted and fixed a few bugs: an overflow bug
in one of the fetcher's comparators, and incorrect handling of
multi-flit messages inside the fetchers.