Set board id in flits emitted by ProgRouter
This was a sutble problem due to the way tinselKeySend works: it uses
a zero address with the isKey bit set as the message destination.
This means that when a message reaches the ProgRouter on a board other
than (0, 0), the board-id of any genenerated local messages is
incorrect.  This could be easily solved in software by tweaking
tinselKeySend, but it's simpler (from a docs/semantics perspective)
and safer to solve it in hardware.