Simplify reliable link
Packet headers inside the reliable link previously consisted of two
64-bit beats.  Here I try using a single 64-bit beat for the header,
and it seems to work fine.

Reading the 10G docs, I thought the Ethertype field (in the second
64-bit beat) was interpreted by the MAC, so couldn't be used for
general data transfer.  But the fact that this change seems to work
suggests it isn't (though I only tried a couple of tests).  As well as
reducing overhead, this change simplifes the reliable link, and means
it should work out-of-the-box for any beat size, e.g. 512-bit beats
for a 100G MAC.

I'll leave this change in a branch for now.  It doesn't matter much
for the 10G links on the DE5.

@lihe1991 you might like to use this version of ReliableLink.bsv when
playing with 100G, rather than the one in the master branch.  You
should just be able to change 64 to 512 and it will work.