New BFS-based partitioner
This is faster than METIS and gives similar results for a few graphs
that I've tried.  It works by picking an unvisited vertex, and then
doing a size-bounded BFS from that vertex with the size-bound equal
to the partition size.  This process is repeated until the partition
is full, at which point we move to a new parition and repeat the
process until all vertices have been visited.  It is just a few lines
of code, and could be easily parallelised in future.