Optimise for multiple dest devices on same thread
When multiple destinations in a pin lie on the same thread, send a
single message to that thread and expand to individual devices
at the receiver.

This is a very basic form of multicasting that's easy to
implement with unicast primitives.  Thus, it provides a better
comparison point for evaluting hardware multicast primitives.