hostlink : retry socket connections + valgrind tweaks + flushing
When running experiments in sequence, sometimes the
next program will fail to open socket the first time (maybe
the daemon is not listening again). This allows a number of
retries for each board before giving up.

valgrind complains about un-init memory problems sent to syscall

flushes file when sending to stdout, so that data gets through to any
consuming process.