Shuffle port A onto valid B locations.
For arb width A,B where A is a mult of B where we want a true dual port the constraints are a bit tricky.

First, we need to pack B's onto A's to match the requirements for equal port widths. Second, we need to pad B's individually to a multiple of 8, so they can be individually written using the byte enables.

With this, we now need to shuffle A's such that the storage exactly overlaps the storage locations of B - so there's a static reordering on the read and write sides.